requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.12
supabase==1.0.3
python-dotenv==1.0.0
streamlit==1.29.0
tiktoken==0.5.2
requests==2.31.0
pypdf==3.17.4